+++ VSIX/RapidXaml.AutoFix/RapidXaml.AutoFix.csproj
@@ -2,7 +2,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
-    <Version>0.12.1</Version>
+    <Version>1.0.3</Version>
   </PropertyGroup>
   <!-- Pack settings -->
   <PropertyGroup>
