+++ VSIX/RapidXamlToolkit.Everything.sln
@@ -63,6 +63,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RapidXaml.Utils", "RapidXam
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RapidXaml.EditorExtras", "RapidXaml.EditorExtras\RapidXaml.EditorExtras.csproj", "{A40BF377-38E3-416B-B25A-5AF0300C1B14}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RapidXaml.Resources", "RapidXaml.Resources\RapidXaml.Resources.csproj", "{068ABBD2-FEBD-4CA1-8B85-5D980DA7BB10}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -476,6 +478,26 @@ Global
 		{A40BF377-38E3-416B-B25A-5AF0300C1B14}.Release|x64.Build.0 = Release|Any CPU
 		{A40BF377-38E3-416B-B25A-5AF0300C1B14}.Release|x86.ActiveCfg = Release|Any CPU
 		{A40BF377-38E3-416B-B25A-5AF0300C1B14}.Release|x86.Build.0 = Release|Any CPU
+		{068ABBD2-FEBD-4CA1-8B85-5D980DA7BB10}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{068ABBD2-FEBD-4CA1-8B85-5D980DA7BB10}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{068ABBD2-FEBD-4CA1-8B85-5D980DA7BB10}.Debug|ARM.ActiveCfg = Debug|Any CPU
+		{068ABBD2-FEBD-4CA1-8B85-5D980DA7BB10}.Debug|ARM.Build.0 = Debug|Any CPU
+		{068ABBD2-FEBD-4CA1-8B85-5D980DA7BB10}.Debug|ARM64.ActiveCfg = Debug|Any CPU
+		{068ABBD2-FEBD-4CA1-8B85-5D980DA7BB10}.Debug|ARM64.Build.0 = Debug|Any CPU
+		{068ABBD2-FEBD-4CA1-8B85-5D980DA7BB10}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{068ABBD2-FEBD-4CA1-8B85-5D980DA7BB10}.Debug|x64.Build.0 = Debug|Any CPU
+		{068ABBD2-FEBD-4CA1-8B85-5D980DA7BB10}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{068ABBD2-FEBD-4CA1-8B85-5D980DA7BB10}.Debug|x86.Build.0 = Debug|Any CPU
+		{068ABBD2-FEBD-4CA1-8B85-5D980DA7BB10}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{068ABBD2-FEBD-4CA1-8B85-5D980DA7BB10}.Release|Any CPU.Build.0 = Release|Any CPU
+		{068ABBD2-FEBD-4CA1-8B85-5D980DA7BB10}.Release|ARM.ActiveCfg = Release|Any CPU
+		{068ABBD2-FEBD-4CA1-8B85-5D980DA7BB10}.Release|ARM.Build.0 = Release|Any CPU
+		{068ABBD2-FEBD-4CA1-8B85-5D980DA7BB10}.Release|ARM64.ActiveCfg = Release|Any CPU
+		{068ABBD2-FEBD-4CA1-8B85-5D980DA7BB10}.Release|ARM64.Build.0 = Release|Any CPU
+		{068ABBD2-FEBD-4CA1-8B85-5D980DA7BB10}.Release|x64.ActiveCfg = Release|Any CPU
+		{068ABBD2-FEBD-4CA1-8B85-5D980DA7BB10}.Release|x64.Build.0 = Release|Any CPU
+		{068ABBD2-FEBD-4CA1-8B85-5D980DA7BB10}.Release|x86.ActiveCfg = Release|Any CPU
+		{068ABBD2-FEBD-4CA1-8B85-5D980DA7BB10}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
