+++ VSIX/RapidXamlToolkit.Tests/RapidXamlToolkit.Tests.csproj
@@ -60,12 +60,15 @@
     <Compile Include="AutoFix\AddAttributeTests.cs" />
     <Compile Include="AutoFix\AddChildStringTests.cs" />
     <Compile Include="AutoFix\AddXmlnsTests.cs" />
+    <Compile Include="AutoFix\BespokeTestFileSystem.cs" />
     <Compile Include="AutoFix\CombinationTests.cs" />
     <Compile Include="AutoFix\ProjectTests.cs" />
     <Compile Include="AutoFix\RemoveAttributeTests.cs" />
     <Compile Include="AutoFix\RemoveChildTests.cs" />
     <Compile Include="AutoFix\RenameElementTests.cs" />
     <Compile Include="AutoFix\ReplaceElementTests.cs" />
+    <Compile Include="AutoFix\SingleFileTests.cs" />
+    <Compile Include="AutoFix\WebViewMultipleActionsAnalyzer.cs" />
     <Compile Include="Extensions\ProjectTypeExtensionTests.cs" />
     <Compile Include="FakeTextSnapshot.cs" />
     <Compile Include="FakeTextSnapshotLine.cs" />
